pyproject.toml
[project]
name = "legacy-flight-autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.11.2",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tavily>=1.1.0",
]
